- Started a major rewrite of the matplotlib backend.  Everything get modularized, which will hopefully enable us (me?) to update parts of the plot selectively.  To be precise: one can say: update only the legend.  Or: update only a certain line.  There is even the possibility to take into account things like "If I change the line, I have to update the legend as well."
- sloppyobject.py is a sample implementation for a new kind of
HasProps container with additional 'parent' property.  But since this
does not work well for TypedList and TypedDict items, I might as well
skip this.

- Default values will be replaced by an entry in the appropriate prop,
which will refer to const.py.  This is not working for pList and
pDictionary right now, because default and reset are keyword arguments
passed on to the TypedList, TypedDict.