I finally realized that this whole Property thing is slowing down my real work: I did not get anything productive done during the last month because I wanted to improve my Property library more and more.
As a consequence, I removed the mapping functionality (which was implemented
in a most horrible way) and restrict the Property classes now to some
kind of 'basic check and value coercing'.  If you want to do fancy stuff,
then it most certainly makes sense to do so in the main loop.

The line treeview in the edit dialog has now been replaced by the new,
automatically created treeview (see widget_factory.py). Some things
are not yet working: (1) colors, (2) values of None, (3) FloatRange
have a default value of Undefined.